refactor world records table into component
